+++ examples/anthropic/script_generator_haiku.py
@@ -1,11 +1,10 @@
 """ 
-Basic example of scraping pipeline using SmartScraper
+Basic example of scraping pipeline using ScriptCreatorGraph
 """
 
 import os
-import json
 from dotenv import load_dotenv
-from scrapegraphai.graphs import SmartScraperGraph
+from scrapegraphai.graphs import ScriptCreatorGraph
 from scrapegraphai.utils import prettify_exec_info
 
 load_dotenv()
@@ -14,32 +13,31 @@ load_dotenv()
 # Define the configuration for the graph
 # ************************************************
 
-
 graph_config = {
     "llm": {
-        "api_key": os.getenv("OPENAI_API_KEY"),
-        "model": "gpt-4o",
+        "api_key": os.getenv("ANTHROPIC_API_KEY"),
+        "model": "anthropic/claude-3-haiku-20240307",
     },
-    "verbose": True,
-    "headless": False,
 }
 
 # ************************************************
-# Create the SmartScraperGraph instance and run it
+# Create the ScriptCreatorGraph instance and run it
 # ************************************************
 
-smart_scraper_graph = SmartScraperGraph(
-    prompt="List me what does the company do, the name and a contact email.",
-    source="https://scrapegraphai.com/",
+script_creator_graph = ScriptCreatorGraph(
+    prompt="List me all the projects with their description.",
+    # also accepts a string with the already downloaded HTML code
+    source="https://perinim.github.io/projects",
     config=graph_config
 )
 
-result = smart_scraper_graph.run()
-print(json.dumps(result, indent=4))
+result = script_creator_graph.run()
+print(result)
 
 # ************************************************
 # Get graph execution info
 # ************************************************
 
-graph_exec_info = smart_scraper_graph.get_execution_info()
+graph_exec_info = script_creator_graph.get_execution_info()
 print(prettify_exec_info(graph_exec_info))
+