+++ examples/anthropic/script_multi_generator_haiku.py
@@ -17,9 +17,8 @@ graph_config = {
     "llm": {
         "api_key": os.getenv("ANTHROPIC_API_KEY"),
         "model": "anthropic/claude-3-haiku-20240307",
-        "max_tokens": 4000
-        },
-        "library": "beautifulsoup"
+    },
+    "library": "beautifulsoup"
 }
 
 # ************************************************