+++ examples/anthropic/search_link_graph_haiku.py
@@ -1,44 +1,45 @@
-""" 
-Basic example of scraping pipeline using ScriptCreatorGraph
 """
-
+Example of Search Graph
+"""
 import os
 from dotenv import load_dotenv
-from scrapegraphai.graphs import ScriptCreatorGraph
-from scrapegraphai.utils import prettify_exec_info
-
-load_dotenv()
+from scrapegraphai.graphs import SearchGraph
+from scrapegraphai.utils import convert_to_csv, convert_to_json, prettify_exec_info
+from langchain_openai import AzureChatOpenAI
+from langchain_openai import AzureOpenAIEmbeddings
 
 # ************************************************
 # Define the configuration for the graph
 # ************************************************
 
+load_dotenv()
+
 graph_config = {
     "llm": {
         "api_key": os.getenv("ANTHROPIC_API_KEY"),
         "model": "anthropic/claude-3-haiku-20240307",
-        "max_tokens": 4000
-        },
+    },
 }
 
 # ************************************************
-# Create the ScriptCreatorGraph instance and run it
+# Create the SearchGraph instance and run it
 # ************************************************
 
-script_creator_graph = ScriptCreatorGraph(
-    prompt="List me all the projects with their description.",
-    # also accepts a string with the already downloaded HTML code
-    source="https://perinim.github.io/projects",
+search_graph = SearchGraph(
+    prompt="List me the best escursions near Trento",
     config=graph_config
 )
 
-result = script_creator_graph.run()
+result = search_graph.run()
 print(result)
 
 # ************************************************
 # Get graph execution info
 # ************************************************
 
-graph_exec_info = script_creator_graph.get_execution_info()
+graph_exec_info = search_graph.get_execution_info()
 print(prettify_exec_info(graph_exec_info))
 
+# Save to json and csv
+convert_to_csv(result, "result")
+convert_to_json(result, "result")