+++ examples/anthropic/xml_scraper_graph_multi_haiku.py
@@ -1,18 +1,18 @@
 """
-Basic example of scraping pipeline using JSONScraperGraph from JSON documents
+Basic example of scraping pipeline using XMLScraperMultiGraph from XML documents
 """
 
 import os
 from dotenv import load_dotenv
-from scrapegraphai.graphs import JSONScraperGraph
+from scrapegraphai.graphs import XMLScraperMultiGraph
 from scrapegraphai.utils import convert_to_csv, convert_to_json, prettify_exec_info
 load_dotenv()
 
 # ************************************************
-# Read the JSON file
+# Read the XML file
 # ************************************************
 
-FILE_NAME = "inputs/example.json"
+FILE_NAME = "inputs/books.xml"
 curr_dir = os.path.dirname(os.path.realpath(__file__))
 file_path = os.path.join(curr_dir, FILE_NAME)
 
@@ -27,31 +27,29 @@ graph_config = {
     "llm": {
         "api_key": os.getenv("ANTHROPIC_API_KEY"),
         "model": "anthropic/claude-3-haiku-20240307",
-        "max_tokens": 4000
-        },
+    },
 }
 
 # ************************************************
-# Create the JSONScraperGraph instance and run it
+# Create the XMLScraperMultiGraph instance and run it
 # ************************************************
 
-json_scraper_graph = JSONScraperGraph(
+xml_scraper_graph = XMLScraperMultiGraph(
     prompt="List me all the authors, title and genres of the books",
-    source=text,  # Pass the content of the file, not the file object
+    source=[text, text],  # Pass the content of the file, not the file object
     config=graph_config
 )
 
-result = json_scraper_graph.run()
+result = xml_scraper_graph.run()
 print(result)
 
 # ************************************************
 # Get graph execution info
 # ************************************************
 
-graph_exec_info = json_scraper_graph.get_execution_info()
+graph_exec_info = xml_scraper_graph.get_execution_info()
 print(prettify_exec_info(graph_exec_info))
 
 # Save to json or csv
 convert_to_csv(result, "result")
 convert_to_json(result, "result")
-