+++ examples/ernie/pdf_scraper_graph_ernie.py
@@ -7,16 +7,12 @@ from scrapegraphai.graphs import PDFScraperGraph
 
 graph_config = {  
     "llm": {
-            "model": "ernie-bot-turbo",
-            "ernie_client_id": "<ernie_client_id>",
-            "ernie_client_secret": "<ernie_client_secret>",
-            "temperature": 0.1
-        },
-        "embeddings": {
-            "model": "ollama/nomic-embed-text",
-            "temperature": 0,
-            "base_url": "http://localhost:11434",}
+        "model": "ernie/ernie-bot-turbo",
+        "ernie_client_id": "<ernie_client_id>",
+        "ernie_client_secret": "<ernie_client_secret>",
+        "temperature": 0.1
     }
+}
 
 source = """
     The Divine Comedy, Italian La Divina Commedia, original name La commedia, long narrative poem written in Italian