+++ examples/ernie/smart_scraper_multi_ernie.py
@@ -12,12 +12,12 @@ load_dotenv()
 # Define the configuration for the graph
 # ************************************************
 
-openai_key = os.getenv("OPENAI_APIKEY")
-
-graph_config = {
+graph_config = {  
     "llm": {
-        "api_key": openai_key,
-        "model": "gpt-4o",
+        "model": "ernie/ernie-bot-turbo",
+        "ernie_client_id": "<ernie_client_id>",
+        "ernie_client_secret": "<ernie_client_secret>",
+        "temperature": 0.1
     },
     "verbose": True,
     "headless": False,