+++ examples/ernie/smart_scraper_schema_ernie.py
@@ -34,10 +34,12 @@ class Projects(BaseModel):
 
 openai_key = os.getenv("OPENAI_APIKEY")
 
-graph_config = {
+graph_config = {  
     "llm": {
-        "api_key": openai_key,
-        "model": "gpt-3.5-turbo",
+        "model": "ernie/ernie-bot-turbo",
+        "ernie_client_id": "<ernie_client_id>",
+        "ernie_client_secret": "<ernie_client_secret>",
+        "temperature": 0.1
     },
     "verbose": True,
     "headless": False,