+++ scrapegraphai/models/oneapi.py
@@ -13,5 +13,7 @@ class OneApi(ChatOpenAI):
         llm_config (dict): Configuration parameters for the language model.
     """
 
-    def __init__(self, llm_config: dict):
+    def __init__(self, **llm_config):
+        if 'api_key' in llm_config:
+            llm_config['openai_api_key'] = llm_config.pop('api_key') 
         super().__init__(**llm_config)